Tests update

+++ tests/TokenManager/tokenManagerTest.php
@@ -7,13 +7,14 @@ class tokenManagerTest extends PHPUnit_Framework_TestCase{
     
     public function __construct($name = null, array $data = array(), $dataName = '') {
         parent::__construct($name, $data, $dataName);
-
-        mkdir($this->tokenDirPath);
+        if(!is_dir($this->tokenDirPath))mkdir($this->tokenDirPath);
     }
     
     public function __destruct() {
-        $this->cleanTokenDir($this->tokenDirPath);
-        rmdir($this->tokenDirPath);
+        if(is_dir($this->tokenDirPath)){
+            $this->cleanTokenDir($this->tokenDirPath);
+            rmdir($this->tokenDirPath);
+        }     
     }
     
     /**
